services/orders/requirements.txt
Flask
psycopg2-binary # Driver para PostgreSQL
python-dotenv   # Para cargar variables de entorno (opcional pero recomendado)